backend/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
pymysql>=1.1
openpyxl>=3.1

